context successfully provided to LandingPage darkMode
