+++ app/(auth)/login/page.jsx
@@ -1,13 +1,24 @@
 'use client'
 import React from "react";
 import Login from "@/app/Components/Login.jsx";
-import { loginProvider } from '@/app/context/Context.jsx';
+import { useAppContext } from "@/app/context";
+import { AppWrapper } from "@/app/context";
 
 const mainLogin = () => {
+    const { 
+		cohorts, 
+		setCohorts,
+		loggedInRole,
+		setLoggedInRole,
+
+	} = useAppContext();
+
     return (
-        <loginProvider>
-            <Login />
-        </loginProvider>
+        <>
+            <AppWrapper>
+                <Login />
+            </AppWrapper>
+        </>
     )
 }
