+++ app/(dashboard)/instructor/page.jsx
@@ -1,6 +1,6 @@
 'use client'
 import React from "react";
-import { UserProvider } from "../../Context/Context";
+import { UserProvider } from "../../context";
 import Instructor from "@/app/Components/InstructorView/Instructor";
 
 export default function InstructorDashboard() {
