--- app/Context/loginProvider.jsx
@@ -1,16 +1,0 @@
-export function loginProvider({ children }) {
-    const [role, setRole] = useState('');
-
-    const changeRole = (id) => {
-        setRole(id);
-    }
-
-    return (
-        <AppContext.Provider value={{
-            role,
-            changeRole
-        }}>
-            {children}
-        </AppContext.Provider>
-    );
-}
