+++ app/layout.js
@@ -1,6 +1,8 @@
 import { Inter } from "next/font/google";
 import "./globals.css";
 import Link from "next/link";
+import { AppWrapper } from "./context";
+
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
@@ -12,13 +14,15 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body>
-      {/* <div className="navbar">
-        <Link href="/">Home</Link>
-        <Link href="/api/students">Student</Link>
-        <Link href="/submit">Submit</Link>
-      </div> */}
-        {children}
-        </body>
+        <AppWrapper>
+          {/* <div className="navbar">
+            <Link href="/">Home</Link>
+            <Link href="/api/students">Student</Link>
+            <Link href="/submit">Submit</Link>
+          </div> */}
+          {children}
+        </AppWrapper>
+      </body>
     </html>
   );
 }
